better modularized user creation; can now create users inside of buying a drink
